requirements.txt
cmlapi>=0.1.0
pyyaml>=6.0
python-dotenv>=0.20.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0 